test: 테스트 위치 변경
